fix: the pipeline's double dispatch or missing dispatch
Add UART controller to the bus. Add the right address of the coming
instruction to get the right target of the branch and jump. Add the
special forwarding for the continuous two load instruction. If the
last inst. is load, insert a bubble and forwarding at the mem stage
only. When the pipeline has continuous two or more stalls, the id
cannot correctly load the new instruction and decode it. So we intro
the neg_inst which will not be the NOP inst but the inst at fetch
stage when there are more than two stalls and the instruction is not
the same as the instruction dispatched previously.